Update README.md
This project demostrates the usage of loops and swtich statements. This program will ask for an input for a different selection of math problems. The user will be given 5 randomly generated problems for the type of problems they want to attemp. After they attemp the selected problems, they are brought back to the input menu, allowing them to practice more problems, or select to exit the program. If they select to exit the program they are given the amount of answer they got correct and the amount attempted and a percentage of problems they got correct.